Refactor processor extrapolation module
