+++ utils/processor_extrapolation_constants.py
@@ -1,0 +1,4 @@
+"""Constants for processor extrapolation utilities."""
+
+MIN_HISTORICAL_DATA_POINTS = 2
+DEFAULT_LOOKBACK_YEARS = 5
